requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
coloredlogs==15.0.1
fastapi==0.116.1
flatbuffers==25.2.10
greenlet==3.2.3
h11==0.16.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
protobuf==6.31.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
